fix(extension): resolve promise for any status code
BREAKING CHANGE: fix the Apiary Chrome extension request() to resolve the returned promise
for any HTTP status code responses. This is the behavior of HTML5 fetch() as well as of the
src/JsAgent/JsAgent.js seed implementation.

Note that the https://github.com/axios/axios client needs to be configured
explicitly for this with the validateStatus functor returning true
otherwise it rejects non-2xx status code requests (see
https://github.com/axios/axios/blob/a105872c1ee2ce3f71ff84a07d9fc27e161a37b0/lib/defaults.js#L79)