fix(extension): resolve promise for any status code
Change the Apiary Chrome extension request() behavior
so that it resolves its promise for any HTTP status code responses.
This is the HTML5 fetch() spec behavior.

Note that the https://github.com/axios/axios client needs to be configured
explicitly for this with the validateStatus functor returning true
otherwise it rejects status code 3xx++ requests.